chore: Extract exception translation to doctest/parts/exception_translator.h
